fix(backend, tests): move prefect class scoped fixtures
Within root conftest.
